(#28) Fix early error about unresolvable extension point
This adds a general facility to report errors that have more than one
source (in this case: the configuration that's missing the extension
point and the method that wants to consume it). The error location is
the configuration class or mount method, the consumer method is
reported as origin.

Also adds navigation to origin in IDEA plugin.