(http build) Add java.time Gson codecs, transitive mvn dependencies
Adds Gson codecs for java.time by default. Default Gson configurations
can now be disabled separately (ServiceLoader, Vavr, java.time).

This introduces a new configuration 'mvnExport' which is extended by 'api'.
This is necessary because the 'implementation' configuration is not mapped
to 'compile' in maven, but 'runtimeOnly', which is not transitive.