(build) Detect detached HEAD on GitInfo and accept as possible release
GitInfo would use the fallback version DEVELOP-SNAPSHOT on CI build on tag,
because CI will checkout the tag, therefore the working tree is in detached
HEAD state, therefore the current branch doesn't match the release candidate
pattern, therefore the release tag on the branch wouldn't be used. Checking
for the branch name HEAD and accepting this as potential release build fixes
this.